Skin switching on weapons complete
