--- src/ServerScriptService/TS/experiment.d.ts
@@ -1,7 +1,0 @@
-declare namespace thing
-{
-    function func1(): number
-    let func2: ()=>number
-}
-
-export = thing
